Change semantics of ivar declarations to not initialize to nil
Instead of initializing declared ivars to nil, they are now kept in a per-class list
of declared ivars. Auditing for unknown ivars now compares prism-analyzed ivars to
the union of instance variables set after instance initialization and declared ivars.

This list of declared ivars functions correctly with inheritance. Declarations in
child classes add to the ivar names declared in superclasses or modules.

-- Auggie